Спринт 7: Перенести метод main в FileBackedTaskManager. Убрать излишнюю проверку на наличие файла в методе save() Удалить класс Main
